fix: resolve all clippy warnings (dead_code, unnecessary_cast, redundant_closure, cmp_owned)

+++ crates/frontend/src/components/common/date_utils.rs
@@ -11,12 +11,8 @@ pub use kartoteka_shared::date_utils::{
 pub fn get_today_string() -> String {
     let d = js_sys::Date::new_0();
     kartoteka_shared::date_utils::format_date(
-        &NaiveDate::from_ymd_opt(
-            d.get_full_year() as i32,
-            d.get_month() as u32 + 1,
-            d.get_date() as u32,
-        )
-        .unwrap_or_default(),
+        &NaiveDate::from_ymd_opt(d.get_full_year() as i32, d.get_month() + 1, d.get_date())
+            .unwrap_or_default(),
     )
 }
